feat(plugin): allow to dispose of a HAR file
To stop the ongoing recording of network requests and disposes of the recorded logs, which will be not saved to a HAR file, new `disposeOfHar` command has been introduced.

You may use this command if you have started recording network logs with `recordHar` command, but you decide to refuse to save the recorded logs to a HAR file.
Here's an example of how you might use this command to dispose of the HAR:

```js
describe('my tests', () => {
  before(() => {
    // start recording
    cy.recordHar();
  });

  after(() => {
    // decide not to save the recorded logs
    cy.disposeOfHar();
  });
});
```

closes #103