feat(plugin):  wait for all requests before saving the HAR file (#148)
You can pass the `waitForIdle` option to wait for all pending requests
to complete before saving the HAR file:

```js
cy.saveHar({ waitForIdle: true });
```

This option is false by default. When set to true, the plugin will
monitor the count of pending requests and wait for it to reach zero
before proceeding with saving the HAR file. This ensures that all
responses have been received and the data in the file is complete and
accurate.

closes #147