feat(plugin): allow filtering entries from HAR using a custom predicate (#172)
You can specify the `filter` option as a path to a JS/TS module that
exports a function to filter out unwanted entries from the HAR. The
function should take an [Entry
object](http://www.softwareishard.com/blog/har-12-spec/#entries) as a
parameter and return a boolean indicating whether the entry should be
included in the final HAR or not.

Here's an example of how to use the `filter` option:

```js
cy.recordHar({ filter: '../support/include-password.ts' });
```

And here's an example of what the `include-password.ts` filter module
might look like:

```ts
import { Entry } from 'har-format';

export default async (req: Entry) => {
  try {
    return /\"password":/.test(req.request.postData.text ?? '');
  } catch {
    return false;
  }
};
```

In this example, the `filter` function will only exclude entries in the
HAR where the request body contains a JSON object with a password field.

You can also specify a `rootDir` option that will be used to resolve the
path of the `filter` option. By default, the path is relative to the
spec folder. But by providing a `rootDir` it will look for the module in
the provided directory:

```js
cy.recordHar({
  filter: 'cypress/support/include-password.ts',
  rootDir: Cypress.config('projectRoot')
});
```

closes #163