feat(plugin): allow controlling the time to wait for idle network (#186)
You can pass the `waitForIdle` option to wait for all pending requests
to complete before saving the HAR file:

```js
cy.saveHar({ waitForIdle: true });
```

This option is false by default. When set to true, the plugin will
monitor the count of pending requests and wait for it to reach zero
before proceeding with saving the HAR file. This ensures that all
responses have been received and the data in the file is complete and
accurate.

Additionally, you can pass the `maxWaitDuration` option to specify the
maximum time to wait for the pending requests to complete:

```js
cy.saveHar({ waitForIdle: true, maxWaitDuration: 20000 });
```

The `maxWaitDuration` option is set to 5000 milliseconds by default,
meaning it will wait for 5 seconds until all pending requests have
completed.

You can also pass the `minIdleDuration` option to specify the minimum
duration in milliseconds to wait for the network idle during the
`maxWaitDuration` time. The network is idle if there are no pending requests during this
time.

```js
cy.saveHar({ waitForIdle: true, minIdleDuration: 1000 });
```

The `minIdleDuration` option is set to 100 milliseconds by default.

closes #185