feat(plugin): make plugin a "no-op" in unsupported browsers
Previously the plugin would throw an error when it was running in an
unsupported browser. This was not great for plugin consumers who run
their cypress tests in both supported and unsupported browsers and are
fine with only having HAR recording in supported browsers: It forced
them to conditionally enable this plugin.

To address this, we're changing the plugin to do nothing except `warn`
that HAR recording is not supported when running in an unsupported
browser.